requirements/test.txt
# SHA1:48eccebd437e7009b0f14f822126b1d3a53770a8
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
attrs==21.4.0
    # via pytest
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
coverage[toml]==6.3.2
    # via pytest-cov
execnet==1.9.0
    # via pytest-xdist
idna==3.3
    # via requests
more-itertools==8.12.0
    # via pytest
ncbi-pytest-selenium==2.16.6
    # via -r requirements/test.in
packaging==21.3
    # via pytest
pillow==9.0.1
    # via ncbi-pytest-selenium
pluggy==0.13.1
    # via pytest
psutil==5.9.0
    # via ncbi-pytest-selenium
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pyparsing==3.0.7
    # via packaging
pytest==5.4.3
    # via
    #   -r requirements/test.in
    #   ncbi-pytest-selenium
    #   pytest-base-url
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-html
    #   pytest-metadata
    #   pytest-pythonpath
    #   pytest-selenium
    #   pytest-variables
    #   pytest-xdist
pytest-base-url==1.4.2
    # via pytest-selenium
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-html==3.1.1
    # via pytest-selenium
pytest-metadata==1.11.0
    # via pytest-html
pytest-pythonpath==0.7.4
    # via -r requirements/test.in
pytest-selenium==1.17.0
    # via ncbi-pytest-selenium
pytest-variables==1.9.0
    # via pytest-selenium
pytest-xdist==1.34.0
    # via ncbi-pytest-selenium
requests==2.27.1
    # via
    #   pytest-base-url
    #   pytest-selenium
selenium==3.141.0
    # via
    #   ncbi-pytest-selenium
    #   pytest-selenium
six==1.16.0
    # via pytest-xdist
teamcity-messages==1.31
    # via -r requirements/test.in
tomli==1.2.3
    # via
    #   -r requirements/test.in
    #   coverage
urllib3==1.26.8
    # via
    #   requests
    #   selenium
wcwidth==0.2.5
    # via pytest
